python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tk-ui"
version = "0.1.0"
description = "Kanban-style ticket viewer TUI for pi .tickets/ directories"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Volker" },
]
keywords = ["kanban", "ticket", "tui", "textual", "pi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "textual>=0.47.0",
    "pyyaml>=6.0",
]

[project.scripts]
tk-ui = "tk_ui.__main__:main"

[project.urls]
Homepage = "https://github.com/legout/pi-ticket-flow"
Repository = "https://github.com/legout/pi-ticket-flow"

[tool.hatch.build.targets.wheel]
packages = ["tk_ui"]

[tool.hatch.build.targets.sdist]
include = [
    "tk_ui/**/*.py",
    "tk_ui/styles.tcss",
]
